code general moderate improvements
Tried some experimental code in order to process the whole file at once but
 found difficulties counting new lines bytes which are essencial for
 counting new lines and bytes (obviously) correctly, so I dropped the
 code and only left a note. That was not too bad as I found about $IFS again
(how could I have forgotten about it?) and was able to improve script speed
 greatly for counting words with IFS, however it turns out
not to be as reliable as the slower and older method, so I am
reverting most changes. The studying was worth it, though!